*refactor full text search methods
